% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeibullEstimates.R
\name{getWeibullEstimates}
\alias{getWeibullEstimates}
\title{Tranforming Log-Linear Estimates to Weibull Estimates}
\usage{
getWeibullEstimates(fit)
}
\arguments{
\item{fit}{object of class \code{survreg} and \code{distribution = "weibull"}.}
}
\value{
Returns ANOVA tables for both the Weibull representation, \code{WeibullModel}, and log-linear
representation, \code{LogLinearModel}. Note, the log-linear estimates are the same as
returned from \code{survreg}. The variance-covariance matrix for the Weibull estimates
can be extracted from \code{var}.
}
\description{
Tranforming Log-Linear Estimates to Weibull Estimates
}
\examples{
library(weibull)
library(survival)
library(KMsurv)
data(alloauto, packge = "KMsurv")
fit <- survreg(Surv(time, delta) ~ 1, data = alloauto, dist = "weibull", 
subset = (type == 1))
getWeibullEstimates(fit)
}
