% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rweibull2.R
\name{rweibull2}
\alias{rweibull2}
\title{Generating observations from a reparameterized Weibull distribution.}
\usage{
rweibull2(n, alpha = 1, lambda = 1)
}
\arguments{
\item{n}{number of observations. If \code{length(n) > 1}, then the length is taken to be the number required.}

\item{alpha}{a positive-valued scale parameter  (Default is 1)}

\item{lambda}{a positive-valued shape parameter (Default is 1)}
}
\description{
Generating observations from a reparameterized Weibull distribution.
}
\details{
Given the scale parameter, \code{alpha}, and shape parameter, \code{lambda},
the density is given by 
\deqn{f(x|\alpha, \lambda) = \alpha\lambdax^(\alpha-1)exp(-\lambdax^\alpha),}
for \eqn{x > 0}. Note that this parameterization is different than the one provided in \code{rweibull}.
}
\examples{
library(weibull)
x <- rweibull2(10000, alpha = 2, lambda = 5)
mean(x)
gamma(1 + 1 / 2) / 5^(1 / 2)
}
