% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeibullDiagnostics.R
\name{getWeibullDiagnostics}
\alias{getWeibullDiagnostics}
\title{Graphical Diagnostic Checks for Weibull AFT Model}
\usage{
getWeibullDiagnostics(fit)
}
\arguments{
\item{fit}{object of class \code{survreg} and \code{distribution = "weibull"}.}
}
\value{
Returns the Cox-Snell residual plot and a plot of the deviance residuals against the study time.
Residuals (Cox-Snell, Martingale, and deviance) are also returned.
}
\description{
Graphical Diagnostic Checks for Weibull AFT Model
}
\examples{
library(weibull)
library(survival)
library(KMsurv)
data(alloauto, packge = "KMsurv")
fit <- survreg(Surv(time, delta) ~ 1, data = alloauto, dist = "weibull", 
subset = (type == 1))
getWeibullDiagnostics(fit)
}
