% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeibullEstimates.R
\name{getWeibullEstimates}
\alias{getWeibullEstimates}
\title{Tranforming Log-Linear Estimates to Weibull Estimates}
\usage{
getWeibullEstimates(fit)
}
\arguments{
\item{fit}{object of class \code{survreg} and \code{distribution = "weibull"}.}
}
\value{
Returns ANOVA tables for both the Weibull representation, \code{WeibullModel}, and log-linear
representation, \code{LogLinearModel}. Note, the log-linear estimates are the same as
returned from \code{survreg}. The variance-covariance matrix for the Weibull estimates
can be extracted from \code{var}.
}
\description{
Tranforming Log-Linear Estimates to Weibull Estimates
}
\examples{
library(weibull)
library(survival)
y <- rexp(100, rate = 2)
delta <- rbinom(100, 1, prob = 0.8)
fit <- survreg(Surv(y, delta) ~ 1, dist = "weibull")
getWeibullEstimates(fit)
}
